requirements.txt
aiogram>=3.7
aiohttp>=3.9
pydantic>=2.7
python-dotenv>=1.0
